Move TestWorldBboxFilter into dynamic data
